consolidated endtoend foreign key tests into onlineddl_scheduler
