+++ go/test/endtoend/onlineddl/scheduler/onlineddl_scheduler_test.go
@@ -26,6 +26,7 @@ import (
 	"time"
 
 	"vitess.io/vitess/go/mysql"
+	"vitess.io/vitess/go/textutil"
 	"vitess.io/vitess/go/vt/schema"
 	"vitess.io/vitess/go/vt/sqlparser"
 
@@ -34,6 +35,10 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+)
+
+const (
+	anyErrorIndicator = "<any-error-of-any-kind>"
 )
 
 var (
@@ -671,6 +676,189 @@ func TestSchemaChange(t *testing.T) {
 	}
 }
 
+func TestForeignKeys(t *testing.T) {
+	defer cluster.PanicHandler(t)
+
+	var (
+		createStatements = []string{
+			`
+			CREATE TABLE parent_table (
+					id INT NOT NULL,
+					parent_hint_col INT NOT NULL DEFAULT 0,
+					PRIMARY KEY (id)
+			)
+		`,
+			`
+			CREATE TABLE child_table (
+				id INT NOT NULL auto_increment,
+				parent_id INT,
+				child_hint_col INT NOT NULL DEFAULT 0,
+				PRIMARY KEY (id),
+				KEY parent_id_idx (parent_id),
+				CONSTRAINT child_parent_fk FOREIGN KEY (parent_id) REFERENCES parent_table(id) ON DELETE CASCADE
+			)
+		`,
+			`
+			CREATE TABLE child_nofk_table (
+				id INT NOT NULL auto_increment,
+				parent_id INT,
+				child_hint_col INT NOT NULL DEFAULT 0,
+				PRIMARY KEY (id),
+				KEY parent_id_idx (parent_id)
+			)
+		`,
+		}
+		insertStatements = []string{
+			"insert into parent_table (id) values(43)",
+			"insert into child_table (id, parent_id) values(1,43)",
+			"insert into child_table (id, parent_id) values(2,43)",
+			"insert into child_table (id, parent_id) values(3,43)",
+			"insert into child_table (id, parent_id) values(4,43)",
+		}
+		ddlStrategy        = "online --allow-zero-in-date"
+		ddlStrategyAllowFK = ddlStrategy + " --unsafe-allow-foreign-keys"
+	)
+
+	type testCase struct {
+		name             string
+		sql              string
+		allowForeignKeys bool
+		expectHint       string
+	}
+	var testCases = []testCase{
+		{
+			name:             "modify parent, not allowed",
+			sql:              "alter table parent_table engine=innodb",
+			allowForeignKeys: false,
+		},
+		{
+			name:             "modify child, not allowed",
+			sql:              "alter table child_table engine=innodb",
+			allowForeignKeys: false,
+		},
+		{
+			name:             "add foreign key to child, not allowed",
+			sql:              "alter table child_table add CONSTRAINT another_fk FOREIGN KEY (parent_id) REFERENCES parent_table(id) ON DELETE CASCADE",
+			allowForeignKeys: false,
+		},
+		{
+			name:             "add foreign key to table which wasn't a child before, not allowed",
+			sql:              "alter table child_nofk_table add CONSTRAINT new_fk FOREIGN KEY (parent_id) REFERENCES parent_table(id) ON DELETE CASCADE",
+			allowForeignKeys: false,
+		},
+		{
+			// on vanilla MySQL, this migration ends with the child_table referencing the old, original table, and not to the new table now called parent_table.
+			// This is a fundamental foreign key limitation, see https://vitess.io/blog/2021-06-15-online-ddl-why-no-fk/
+			// However, this tests is still valid in the sense that it lets us modify the parent table in the first place.
+			name:             "modify parent, trivial",
+			sql:              "alter table parent_table engine=innodb",
+			allowForeignKeys: true,
+			expectHint:       "parent_hint_col",
+		},
+		{
+			// on vanilla MySQL, this migration ends with two tables, the original and the new child_table, both referencing parent_table. This has
+			// the unwanted property of then limiting actions on the parent_table based on what rows exist or do not exist on the now stale old
+			// child table.
+			// This is a fundamental foreign key limitation, see https://vitess.io/blog/2021-06-15-online-ddl-why-no-fk/
+			// However, this tests is still valid in the sense that it lets us modify the child table in the first place.
+			// A valid use case: using FOREIGN_KEY_CHECKS=0  at all times.
+			name:             "modify child, trivial",
+			sql:              "alter table child_table engine=innodb",
+			allowForeignKeys: true,
+			expectHint:       "REFERENCES `parent_table`",
+		},
+		{
+			// on vanilla MySQL, this migration ends with two tables, the original and the new child_table, both referencing parent_table. This has
+			// the unwanted property of then limiting actions on the parent_table based on what rows exist or do not exist on the now stale old
+			// child table.
+			// This is a fundamental foreign key limitation, see https://vitess.io/blog/2021-06-15-online-ddl-why-no-fk/
+			// However, this tests is still valid in the sense that it lets us modify the child table in the first place.
+			// A valid use case: using FOREIGN_KEY_CHECKS=0  at all times.
+			name:             "add foreign key to child",
+			sql:              "alter table child_table add CONSTRAINT another_fk FOREIGN KEY (parent_id) REFERENCES parent_table(id) ON DELETE CASCADE",
+			allowForeignKeys: true,
+			expectHint:       "another_fk",
+		},
+		{
+			name:             "add foreign key to table which wasn't a child before",
+			sql:              "alter table child_nofk_table add CONSTRAINT new_fk FOREIGN KEY (parent_id) REFERENCES parent_table(id) ON DELETE CASCADE",
+			allowForeignKeys: true,
+			expectHint:       "new_fk",
+		},
+	}
+
+	testStatement := func(t *testing.T, sql string, ddlStrategy string, expectHint string, expectError bool) (uuid string) {
+		errorHint := ""
+		if expectError {
+			errorHint = anyErrorIndicator
+		}
+		return testOnlineDDLStatement(t, sql, ddlStrategy, "vtctl", expectHint, errorHint, false)
+	}
+	for _, testcase := range testCases {
+		t.Run(testcase.name, func(t *testing.T) {
+			t.Run("create tables", func(t *testing.T) {
+				for _, statement := range createStatements {
+					t.Run(statement, func(t *testing.T) {
+						uuid := testStatement(t, statement, ddlStrategyAllowFK, "", false)
+						onlineddl.CheckMigrationStatus(t, &vtParams, shards, uuid, schema.OnlineDDLStatusComplete)
+					})
+				}
+			})
+			t.Run("populate tables", func(t *testing.T) {
+				for _, statement := range insertStatements {
+					t.Run(statement, func(t *testing.T) {
+						onlineddl.VtgateExecQuery(t, &vtParams, statement, "")
+					})
+				}
+			})
+			var uuid string
+			t.Run("run migration", func(t *testing.T) {
+				if testcase.allowForeignKeys {
+					uuid = testStatement(t, testcase.sql, ddlStrategyAllowFK, testcase.expectHint, false)
+					onlineddl.CheckMigrationStatus(t, &vtParams, shards, uuid, schema.OnlineDDLStatusComplete)
+				} else {
+					uuid = testStatement(t, testcase.sql, ddlStrategy, "", true)
+					if uuid != "" {
+						onlineddl.CheckMigrationStatus(t, &vtParams, shards, uuid, schema.OnlineDDLStatusFailed)
+					}
+				}
+			})
+			t.Run("cleanup", func(t *testing.T) {
+				var artifacts []string
+				if uuid != "" {
+					rs := onlineddl.ReadMigrations(t, &vtParams, uuid)
+					require.NotNil(t, rs)
+					row := rs.Named().Row()
+					require.NotNil(t, row)
+
+					artifacts = textutil.SplitDelimitedList(row.AsString("artifacts", ""))
+				}
+
+				artifacts = append(artifacts, "child_table", "child_nofk_table", "parent_table")
+				// brute force drop all tables. In MySQL 8.0 you can do a single `DROP TABLE ... <list of all tables>`
+				// which auto-resovled order. But in 5.7 you can't.
+				droppedTables := map[string]bool{}
+				for range artifacts {
+					for _, artifact := range artifacts {
+						if droppedTables[artifact] {
+							continue
+						}
+						statement := fmt.Sprintf("DROP TABLE IF EXISTS %s", artifact)
+						_, err := clusterInstance.VtctlclientProcess.ApplySchemaWithOutput(keyspaceName, statement, cluster.VtctlClientParams{DDLStrategy: "direct", SkipPreflight: true})
+						if err == nil {
+							droppedTables[artifact] = true
+						}
+					}
+				}
+				statement := fmt.Sprintf("DROP TABLE IF EXISTS %s", strings.Join(artifacts, ","))
+				t.Run(statement, func(t *testing.T) {
+					testStatement(t, statement, "direct", "", false)
+				})
+			})
+		})
+	}
+}
+
 // testOnlineDDLStatement runs an online DDL, ALTER statement
 func testOnlineDDLStatement(t *testing.T, ddlStatement string, ddlStrategy string, executeStrategy string, expectHint string, expectError string, skipWait bool) (uuid string) {
 	strategySetting, err := schema.ParseDDLStrategy(ddlStrategy)
@@ -696,10 +884,18 @@ func testOnlineDDLStatement(t *testing.T, ddlStatement string, ddlStrategy strin
 			params = overrideVtctlParams
 		}
 		output, err := clusterInstance.VtctlclientProcess.ApplySchemaWithOutput(keyspaceName, ddlStatement, *params)
-		if expectError == "" {
+		switch expectError {
+		case anyErrorIndicator:
+			if err != nil {
+				// fine. We got any error.
+				t.Logf("expected any error, got this error: %v", err)
+				return
+			}
+			uuid = output
+		case "":
 			assert.NoError(t, err)
 			uuid = output
-		} else {
+		default:
 			assert.Error(t, err)
 			assert.Contains(t, output, expectError)
 		}
